refactor: Moved styling logic to `AntStyledSelect`
